Move packages around and use namespace nkms.crypto.keypairs and nkms.crypto.keyring
